Squashed commit of the following:

commit 898c43612ae918d39a1f20a6bbb9c1d0e2917c11
Author: Victor Roemer <viroemer@cisco.com>
Date:   Wed Dec 9 13:32:41 2015 -0500

    Remove global is_two_way_traffic.

commit d923c47d4b25b098ddceffaf78bba2d5229852a7
Author: Victor Roemer <viroemer@cisco.com>
Date:   Tue Dec 8 19:01:11 2015 -0500

    Uncrustify

commit 84131aa41ee3e44c0f52ec186b4275a5c2f493d5
Author: Victor Roemer <viroemer@cisco.com>
Date:   Tue Dec 8 18:54:31 2015 -0500

    Packets from a passive nic are readonly.

commit a9e68a7dec7821a20a42a3494a9d41f238473e72
Author: Victor Roemer <viroemer@cisco.com>
Date:   Mon Dec 7 16:24:30 2015 -0500

    FIXIT note for broken assert() in nhttp.

commit b042cf28c498faac92148ce10f0182811cff58f4
Author: Victor Roemer <viroemer@cisco.com>
Date:   Fri Dec 4 14:51:26 2015 -0500

    Sync changes up-to Snort 2.9.7.5.

+++ src/service_inspectors/http_inspect/hi_main.cc
@@ -1085,7 +1085,8 @@ int HttpInspectMain(HTTPINSPECT_CONF* conf, Packet* p)
                 }
                 else
                 {
-                    set_file_data((uint8_t*)session->server.response.body, detect_data_size);
+                    set_file_data((uint8_t*)session->server.response.body,
+                        (uint16_t)detect_data_size);
                 }
 
                 FileFlows* file_flows = FileFlows::get_file_flows(p->flow);
